Remove obsolete generators and files
Remove the gapic and gapic_dump generators, templates, and tests.
Remove the library-based helpers class, use the template helpers instead.
