+++ templates/default/client.erb
@@ -1,8 +1,7 @@
-# frozen_string_literal: true
-
-<%= render partial: "shared/license" %>
-
-require "google/gax"
-
-<%= api.class %>
-<%= service.class %>
+<%= render partial: "shared/client",
+           layout: "layouts/ruby",
+           locals: {
+             service: service,
+             namespaces: service.address[0..-2]
+           }
+%>
