+++ templates/default/helpers/namespace_helper.rb
@@ -26,4 +26,11 @@ module NamespaceHelper
       puts "end"
     end
   end
+
+  ##
+  # Converts a dot-separated namespace string to a new string with
+  # Ruby double-semicolon separators.
+  def ruby_namespace str
+    str.split(".").reject(&:empty?).map(&:upcase_first).join "::"
+  end
 end
