+++ templates/default/shared/_client.erb
