--- templates/gapic/clients.erb
@@ -1,11 +1,0 @@
-<% @api.services.each do |service|
-     render_to_file "#{service.name.downcase}.rb" do %>
-<%= render partial: "shared/client",
-           layout: "layouts/ruby",
-           locals: {
-             service: service,
-             namespaces: service.address[0..-2]
-           }
-%>
-<%   end
-   end -%>
