--- templates/gapic_dump/gapic_dump.erb
@@ -1,58 +1,0 @@
-<% render_to_file "gapic_dump.txt" do %>
-<% @api.messages.each do |m| %>
-Message: <%= m.name %>
-  Leading Comments: <%= m.leading_comments.inspect %>
-  Trailing Comments: <%= m.trailing_comments.inspect %>
-  Address: <%= m.address %>
-  Fields:
-  <% m.fields.each do |f| %>
-    Field: <%= f.name %>
-      Leading Comments: <%= f.leading_comments.inspect %>
-      Trailing Comments: <%= f.trailing_comments.inspect %>
-      Address: <%= f.address %>
-    <% if f.message %>
-      Message: <%= f.message.name %>
-    <% elsif f.enum %>
-      Enum: <%= f.enum.name %>
-    <% else %>
-      Type: <%= f.type %>
-    <% end %>
-  <% end %>
-
-<% end %>
-
-<% @api.enums.each do |e| %>
-Enum: <%= e.name %>
-  Leading Comments: <%= e.leading_comments.inspect %>
-  Trailing Comments: <%= e.trailing_comments.inspect %>
-  Address: <%= e.address %>
-  Values:
-  <% e.values.each do |v| %>
-    EnumValue: <%= v.name %>
-      Leading Comments: <%= v.leading_comments.inspect %>
-      Trailing Comments: <%= v.trailing_comments.inspect %>
-      Address: <%= v.address %>
-      Number: <%= v.number %>
-  <% end %>
-
-<% end %>
-
-<% @api.services.each do |s| %>
-Service: <%= s.name %>
-  Leading Comments: <%= s.leading_comments.inspect %>
-  Trailing Comments: <%= s.trailing_comments.inspect %>
-  Address <%= s.address %>
-  <% s.methods.each do |m| %>
-    Method: <%= m.name %>
-      Leading Comments: <%= m.leading_comments.inspect %>
-      Trailing Comments: <%= m.trailing_comments.inspect %>
-      Address <%= m.address %>
-      Input Type: <%= m.input.name %>
-        Address: <%= m.input.address %>
-      Output Type: <%= m.output.name %>
-        Address: <%= m.output.address %>
-
-  <% end %>
-
-<% end %>
-<% end %>
